fix: dispatch each function to different files
