+++ server/src/modules/Utils/hashedPassword.ts
@@ -1,0 +1,23 @@
+import argon2 from "argon2";
+import type { RequestHandler } from "express";
+
+const hashingOptions = {
+  type: argon2.argon2id,
+  memoryCost: 19 * 2 ** 10,
+  timeCost: 2,
+  parallelism: 1,
+};
+
+const hashPassword: RequestHandler = async (req, res, next) => {
+  try {
+    const { password } = req.body;
+    const hashedPassword = await argon2.hash(password, hashingOptions);
+    req.body.hashed_password = hashedPassword;
+    req.body.password = undefined;
+    next();
+  } catch (err) {
+    next(err);
+  }
+};
+
+export default hashPassword;
